backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.1
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dateutil>=2.9.0.post0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5
python-multipart>=0.0.9

